% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_predict.R
\name{percentile_net_worth}
\alias{percentile_net_worth}
\title{Calculate the percentile of the net worth given age and net worth}
\usage{
percentile_net_worth(
  age = NULL,
  net_worth = NULL,
  net_worth_models = shinyVirga::get_userData("net_worth_models")
)
}
\arguments{
\item{age}{\code{num} Age of the person}

\item{net_worth}{\code{num} Net worth of the person}

\item{net_worth_models}{\code{list} Bureau of Labor Statistics models made with ../SCF_Munging.qmd}
}
\value{
\code{list} with `age`, `net_worth`, and `p` (the percentile of the net worth)
}
\description{
Calculate the percentile of the net worth given age and net worth
}
\examples{
percentile_net_worth(age = 39, net_worth = 126000)
}
