% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_predict.R
\name{predict_net_worth}
\alias{predict_net_worth}
\title{Predict the net worth of the p1 at a specific age}
\usage{
predict_net_worth(
  p1,
  age,
  net_worth_models = shinyVirga::get_userData("net_worth_models")
)
}
\arguments{
\item{p1}{\code{list} with variables `age` and one of `net_worth` or `p` (the percentile)}

\item{age}{\code{num} Age at which to predict net worth based on percentile}

\item{net_worth_models}{\code{list} Bureau of Labor Statistics models made with ../SCF_Munging.qmd}
}
\value{
predicted net worth of p1 at the age specified
}
\description{
Predict the net worth of the p1 at a specific age
}
\examples{
predict_net_worth(percentile_net_worth(age = 39, net_worth = 113000), 42)
}
\seealso{
[percentile_net_worth()]
}
