% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_predict.R
\name{sliding_scale_factor}
\alias{sliding_scale_factor}
\title{Derive the sliding scale multiplicative factor if p1 is paid by p2}
\usage{
sliding_scale_factor(
  p1,
  p2,
  net_worth_models = shinyVirga::get_userData("net_worth_models")
)
}
\arguments{
\item{p1}{\code{list} with variables `age` and one of `net_worth` or `p` (the percentile)}

\item{net_worth_models}{\code{list} Bureau of Labor Statistics models made with ../SCF_Munging.qmd}
}
\value{
\code{num} The multiplicative factor for the base rate
}
\description{
Derive the sliding scale multiplicative factor if p1 is paid by p2
}
\examples{
sliding_scale_factor(percentile_net_worth(age = 39, net_worth = 126000, net_worth_models), percentile_net_worth(age = 42, net_worth = 250000, net_worth_models))
}
